# common-platform-customer-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-customer-mgmt
  description: Customer Management Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/customers
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/parties/{partyId}:
    get:
      tags:
        - Parties
      summary: Get party by ID
      description: Retrieve a specific party by its unique identifier
      operationId: getPartyById
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Party found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartyDTO'
        '404':
          description: Party not found
    put:
      tags:
        - Parties
      summary: Update party
      description: Update an existing party with new information
      operationId: updateParty
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyDTO'
        required: true
      responses:
        '200':
          description: Party successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartyDTO'
        '400':
          description: Invalid party data
        '404':
          description: Party not found
    delete:
      tags:
        - Parties
      summary: Delete party
      description: Delete a party from the system
      operationId: deleteParty
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Party successfully deleted
        '404':
          description: Party not found
...
  /api/v1/parties:
    post:
      tags:
        - Parties
      summary: Create party
      description: Create a new party in the system
      operationId: createParty
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyDTO'
        required: true
      responses:
        '201':
          description: Party successfully created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartyDTO'
        '400':
          description: Invalid party data
...
  /api/v1/parties/filter:
    post:
      tags:
        - Parties
      summary: Filter parties
      description: Retrieve a paginated list of parties based on filtering criteria
      operationId: filterParties
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestPartyDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved filtered parties
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
...
components:
  schemas:
    PartyDTO:
      required:
        - partyKind
      type: object
      properties:
        partyId:
          type: string
          format: uuid
          readOnly: true
          example: "123e4567-e89b-12d3-a456-426614174000"
        partyKind:
          type: string
          enum:
            - INDIVIDUAL
            - ORGANIZATION
        preferredLanguage:
          maxLength: 10
          minLength: 0
          type: string
        sourceSystem:
          maxLength: 100
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestPartyDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/PartyDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for parties